Car Profile Page for Android In-Car App
Create a car profile page complete with button to clear current database
+ button ("dummy button") meant to simulate the login button.

This profile page should represent the second screen someone sees in the app.

This page is only displayed once, the first time the user displays the
page.  We will add the feature to skip the form later, but for now,
when you have previously submit the form, it will simply skip this
profile page.